added a bunch of documentation.

+++ lazy_slides/manipulation/convert.py
@@ -7,6 +7,15 @@ from .PythonMagickWand import *
 log = logging.getLogger(__name__)
 
 def convert(infilename, target_type='png'):
+    '''Convert an image from one type to another.
+
+    This uses MagickWand to convert an input file into another
+    type. After conversion, the original file is deleted.
+
+    :param infilename: The name of the file to convert.
+    :param target_type: The new image type to save as.
+    '''
+
     outfilename = '{}.{}'.format(
         os.path.splitext(infilename)[0],
         target_type)
@@ -19,8 +28,13 @@ def convert(infilename, target_type='png'):
     MagickReadImage(wand, infilename.encode('utf-8'))
     MagickWriteImage(wand, outfilename.encode('utf-8'))
 
+    # TODO: Reconsider this. The deletion should probably occur
+    # elsewhere (if anywhere.) Having it here will mess with caching,
+    # etc.
+    #
     # Delete original
-    log.debug('Deleting {}'.format(infilename))
-    os.remove(infilename)
+    if infilename != outfilename:
+        log.debug('Deleting {}'.format(infilename))
+        os.remove(infilename)
 
     return outfilename
